add ErrorBoundary component
to assist with debugging, add an ErrorBoundary component that catches
any error when rendering a child component.
